Cleaned Diagnostic Breast Cancer Data

A subset of a dataset containing quantitative features from images of nuclei from biopsies of breast masses from patients at the University of Wisconsin Hospital. Malignancy is encoded as a binary variable (1 = malignant, 0 = benign).

cancer_clean

Format

A tibble with 569 rows and 11 variables:

malignant

Whether the observation was diagnosed as malignant or benign

radius_mean

Mean radius of nuclei present in sample image

texture_mean

Mean texture of nuclei present in sample image

perimeter_mean

Mean perimeter of nuclei present in sample image

area_mean

Mean area of nuclei present in sample image

smoothness_mean

Mean smoothness of nuclei present in sample image

compactness_mean

Mean compactness of nuclei present in sample image

concavity_mean

Mean concavity of nuclei present in sample image

concave_points_mean

Mean concave points of nuclei present in sample image

symmetry_mean

Mean symmetry of nuclei present in sample image

fractal_dimension_mean

Mean factal dimension of nuclei present in sample image

Source

https://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Diagnostic)