GLM Prediction

Create a Generalised Linear Model (glm) from a dataset and a given set of variables, then make predictions from the model, either on the original data or a new dataset.

glm_predict(model, target, threshold = 0.5, newdata = NULL, ...)

Arguments

model
Fitted glm model object

target
Target name (must be present as a column in the model object). This is the y variable when creating a glm.

threshold
Probability threshold for predicting positive class. Must be between 0 and 1 (exclusive). default=0.5.

newdata
New data to make predictions on. By default, glm_predict predicts on the original data used to create the model. default=NULL.

...
Other parameters to be passed into the augment function for making predictions.

Value

A named list containing:

$augment: Augment object containing the model predictions and information about each observation.

$accuracy: Accuracy of predictions.

$precision: Precision of predictions.

$recall: Recall of predictions.

Examples

require(dplyr)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

cancer_model <- glm(malignant ~ texture_mean, data=cancer_clean, family="quasibinomial")
glm_predict(cancer_model, "malignant")

#> $augment
#> # A tibble: 569 x 9
#>    malignant texture_mean .fitted .resid .std.resid    .hat .sigma .cooksd
#>        <dbl>        <dbl>   <dbl>  <dbl>      <dbl>   <dbl>  <dbl>   <dbl>
#>  1         1         10.4  0.0636  2.35       2.33  0.00415   1.06 0.0304 
#>  2         1         17.8  0.278   1.60       1.59  0.00239   1.07 0.00308
#>  3         1         21.2  0.465   1.24       1.23  0.00263   1.07 0.00149
#>  4         1         20.4  0.415   1.33       1.32  0.00227   1.07 0.00158
#>  5         1         14.3  0.147   1.96       1.95  0.00369   1.07 0.0107 
#>  6         1         15.7  0.191   1.82       1.81  0.00319   1.07 0.00668
#>  7         1         20.0  0.392   1.37       1.36  0.00218   1.07 0.00167
#>  8         1         20.8  0.441   1.28       1.27  0.00243   1.07 0.00152
#>  9         1         21.8  0.499   1.18       1.17  0.00298   1.07 0.00148
#> 10         1         24.0  0.626   0.968      0.963 0.00501   1.07 0.00149
#> # … with 559 more rows, and 1 more variable: .prediction <dbl>
#> 
#> $accuracy
#> [1] 0.7047452
#> 
#> $precision
#> [1] 0.6594203
#> 
#> $recall
#> [1] 0.4292453
#> 

glm_predict(cancer_model, "malignant", threshold=0.8)

#> $augment
#> # A tibble: 569 x 9
#>    malignant texture_mean .fitted .resid .std.resid    .hat .sigma .cooksd
#>        <dbl>        <dbl>   <dbl>  <dbl>      <dbl>   <dbl>  <dbl>   <dbl>
#>  1         1         10.4  0.0636  2.35       2.33  0.00415   1.06 0.0304 
#>  2         1         17.8  0.278   1.60       1.59  0.00239   1.07 0.00308
#>  3         1         21.2  0.465   1.24       1.23  0.00263   1.07 0.00149
#>  4         1         20.4  0.415   1.33       1.32  0.00227   1.07 0.00158
#>  5         1         14.3  0.147   1.96       1.95  0.00369   1.07 0.0107 
#>  6         1         15.7  0.191   1.82       1.81  0.00319   1.07 0.00668
#>  7         1         20.0  0.392   1.37       1.36  0.00218   1.07 0.00167
#>  8         1         20.8  0.441   1.28       1.27  0.00243   1.07 0.00152
#>  9         1         21.8  0.499   1.18       1.17  0.00298   1.07 0.00148
#> 10         1         24.0  0.626   0.968      0.963 0.00501   1.07 0.00149
#> # … with 559 more rows, and 1 more variable: .prediction <dbl>
#> 
#> $accuracy
#> [1] 0.6239016
#> 
#> $precision
#> [1] 0.4615385
#> 
#> $recall
#> [1] 0.05660377
#>